Prune package of non-supported templates
The Heat team's upstream template repository has always served the
purpose of being for example use only. Because of this fact, we want to
package only the set of templates we can support long term. Everything
that was not software-config related has now been removed. However, the
remaining templates will still be available directly from github here:

https://github.com/openstack/heat-templates

The dependencies were removed on all unnecessary packages, including
that of Heat and diskimage-builder given that the templates or elements
may be used separately from one another.