python-heat-agent subpackage for simple agent bootstrap
This change creates a new subpackage python-heat-agent out of
openstack-heat-templates.

Currently image building or boot configuration has to do a number of
non-obvious steps to end up with a server ready to perform heat-driven
software deployments.

The package python-heat-agent installs all dependencies and files
required to do this, resulting in a boot config on a pristine image
being as simple as:

        yum -y install https://www.rdoproject.org/repos/rdo-release.rpm
        yum -y install python-heat-agent
        systemctl enable os-collect-config
        systemctl start --no-block os-collect-config

python-heat-agent installs one hook which allows configuration via
heat templates. The aim is to create another subpackage per
configuration tool hook in heat-templates. So python-heat-agent-puppet
will install the puppet hook and depend on python-heat-agent and
puppet.

This depends on some upstream heat-templates changes:
https://review.openstack.org/#/q/topic:centos-rdo-boot-config

Change-Id: I39025e648876d0e2b6c527b1fed4589f5cbf37b7